# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API de Gestão de Projetos",
        "description": "Documentação da API para o sistema de gestão de projetos de energia solar",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/enderecos": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Listar todos os endereços",
                "description": "Retorna uma lista de todos os endereços.",
                "operationId": "eaa67d6b1d5ab2d3487483bf7cd9ee20",
                "responses": {
                    "200": {
                        "description": "Lista de endereços",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Endereco"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Criar um novo endereço",
                "description": "Cria um novo endereço com os dados fornecidos.",
                "operationId": "7503e124e0fd8144c6a9901d2e08bb19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uf"
                                ],
                                "properties": {
                                    "uf": {
                                        "description": "UF do endereço",
                                        "type": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Endereco"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar endereço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar endereço"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/enderecos/{id}": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Obter detalhes de um endereço",
                "description": "Retorna os detalhes de um endereço específico baseado no ID fornecido.",
                "operationId": "04f205c824b703870cfdbbbcf75cc3f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do endereço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Endereco"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Atualizar um endereço",
                "description": "Atualiza os dados de um endereço existente baseado no ID fornecido.",
                "operationId": "5aac4f6f8753fbad5804363cd0f793f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uf"
                                ],
                                "properties": {
                                    "uf": {
                                        "description": "UF do endereço",
                                        "type": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Endereco"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar endereço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar endereço"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Deletar um endereço",
                "description": "Deleta um endereço existente baseado no ID fornecido.",
                "operationId": "9f4f92c7f07c1b2f09355a725a9d51cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar endereço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao deletar endereço"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Endereco": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uf": {
                        "type": "string",
                        "example": "SP"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}